Drop trailing \b from tag regular expression
For some reason when encountering a non-ASCII character at the end of a tag
the engine would backtrack so the \b can consume it, instead of matching the
next space, punctuation or end of string to the \b. So for example #menü would
get matched as #men.

Just dropping the \b doesn't seem to have any adverse effects.